corrijo simulaciones

+++ src/main/java/com/utn/simulador/negocio/simuladornegocio/domain/EstadoInicial.java
@@ -20,7 +20,7 @@ public class EstadoInicial {
     private Long stock;
     private Long produccionMensual;
     private Integer calidad;
-    private Integer publicidad;
+    private BigDecimal publicidad;
     private Integer cantidadVendedores;
 
 }
